Refactor Background Files
