+++ src/components/shared/background/_background.scss
@@ -1,0 +1,60 @@
+.background-block {
+    position: relative;
+    height: max-content;
+}
+.main-background-1 {
+    position: relative;
+    width: 100vw;
+    min-width: 1365px;
+    height: auto;
+
+    /* Animation Triggering Fix */
+    margin-top: -8vw;
+    padding-top: 5vw;
+}
+.main-background-2 {
+    width: 100vw;
+    min-width: 1365px;
+    height: auto;
+}
+.ghost-three {
+    position: fixed;
+    width: 100vw;
+    min-width: 1365px;
+    height: auto;
+    top: max(750px, 65vw);
+    transform: rotate(5deg);
+}
+@media all and (min-width: 1550px) {
+    .main-background-2 {
+        top: 85vw;
+    }
+}
+@media all and (max-width: 1365px) {
+    .main-background-1,
+    .main-background-2 {
+        transform: translateX(calc(-20% + 20vw));
+        min-width: 800px !important;
+    }
+}
+@media all and (max-width: 780px) {
+    .main-background-1,
+    .main-background-2 {
+        transform: translateX(calc(-25% + 20vw));
+        min-width: 800px !important;
+    }
+}
+@media all and (max-width: 600px) {
+    .main-background-1,
+    .main-background-2 {
+        transform: translateX(calc(-42% + 20vw));
+        min-width: 800px !important;
+    }
+}
+@media all and (max-width: 450px) {
+    .main-background-1,
+    .main-background-2 {
+        transform: translateX(calc(-50% + 20vw));
+        min-width: 800px !important;
+    }
+} ;
